Move env vars to constants
