+++ src/index.ts
@@ -25,16 +25,15 @@ import {
   stopPollingLoLGame
 } from './league-of-legends-api/poll-current-game';
 import { BotCommands } from './types';
+import {
+  PORT,
+  DISCORD_BOT_TOKEN,
+  IS_LOL_ANNOUNCER_ENABLED,
+  PATH_TO_CLIPS,
+  GUILD_ID
+} from './constants';
 
 const app = express();
-
-// Grab environment variables from the .env file
-const PORT = process.env.PORT || 8081;
-const DISCORD_BOT_TOKEN: string = process.env.DISCORD_BOT_TOKEN || '';
-const IS_LOL_ANNOUNCER_ENABLED: boolean =
-  Boolean(process.env.LEAGUE_OF_LEGENDS_ANNOUNCER_ENABLED) ?? false;
-export const PATH_TO_CLIPS: string = process.env.PATH_TO_CLIPS || '';
-const GUILD_ID: string = process.env.GUILD_ID || '';
 
 // Create the bot
 const client = new Client({
